공지사항 시큐리티 수정

+++ BE/src/main/java/com/dignity/puppymarket/controller/NoticeController.java
@@ -2,10 +2,11 @@ package com.dignity.puppymarket.controller;
 
 import com.dignity.puppymarket.dto.NoticeRequestDto;
 import com.dignity.puppymarket.dto.NoticeResponseDto;
+import com.dignity.puppymarket.security.UserAuthentication;
 import com.dignity.puppymarket.service.NoticeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -39,24 +40,36 @@ public class NoticeController {
 
     //공지글 수정
     @PutMapping("/{id}")
-    @PreAuthorize("isAuthenticated() and hasAuthority('ADMIN')")
-    public NoticeResponseDto updateNotice(@PathVariable Long id, @RequestBody NoticeRequestDto form){
+    //@PreAuthorize("isAuthenticated() and hasAuthority('ADMIN')")
+    public NoticeResponseDto updateNotice(@PathVariable Long id, @RequestBody NoticeRequestDto form,
+                                          UserAuthentication userAuthentication){
+        validateAdmin(userAuthentication);
         return noticeService.updateNotice(id, form);
     }
 
     //공지글 저장
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    @PreAuthorize("isAuthenticated() and hasAuthority('ADMIN')")
-    public NoticeResponseDto createNotice(@RequestBody NoticeRequestDto form){
+    //@PreAuthorize("isAuthenticated() and hasAuthority('ADMIN')")
+    public NoticeResponseDto createNotice(@RequestBody NoticeRequestDto form,
+                                          UserAuthentication userAuthentication){
+        validateAdmin(userAuthentication);
         return noticeService.saveNotice(form);
     }
 
     //공지글 삭제
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PreAuthorize("isAuthenticated() and hasAuthority('ADMIN')")
-    public void deleteNotice(@PathVariable Long id) {
+    //@PreAuthorize("isAuthenticated() and hasAuthority('ADMIN')")
+    public void deleteNotice(@PathVariable Long id,
+                             UserAuthentication userAuthentication) {
+        validateAdmin(userAuthentication);
         noticeService.deleteNotice(id);
+    }
+
+    public void validateAdmin(UserAuthentication userAuthentication) {
+        if(userAuthentication == null || !userAuthentication.getEmail().equals("admin")) {
+            throw new AccessDeniedException("권한이 없습니다. 어드민만 접근 가능합니다");
+        }
     }
 }
